Migrate utils RBAC to TS
