+++ packages/utils/package.json
@@ -29,6 +29,7 @@
     "peerDependencies": {
         "@patternfly/react-core": ">=4.18.5",
         "@patternfly/react-table": ">=2.1.6",
+        "@redhat-cloud-services/rbac-client": ">=1.0.100",
         "react": ">=16.14.0 || >=17.0.0",
         "react-dom": ">=16.14.0 || >=17.0.0",
         "react-redux": ">=5.0.7",
