--- packages/utils/src/RBAC/RBAC.js
@@ -1,21 +1,0 @@
-export async function getRBAC(applicationName = '', disableCache = false) {
-  const insights = window.insights;
-  const user = await insights?.chrome?.auth?.getUser();
-  return {
-    // eslint-disable-next-line camelcase
-    isOrgAdmin: user?.identity?.user?.is_org_admin || false,
-    permissions: (await insights?.chrome?.getUserPermissions(applicationName, disableCache)) || null,
-  };
-}
-
-export function doesHavePermissions(userPermissions, permissionList) {
-  if (!userPermissions) {
-    return false;
-  }
-
-  return userPermissions.some((access) => {
-    return permissionList.includes(access?.permission || access);
-  });
-}
-
-export default getRBAC;
