--- packages/utils/src/RBAC/index.js
@@ -1,2 +1,0 @@
-export { default } from './RBAC';
-export * from './RBAC';
